requirements_venv_chat_backend_test.txt
elasticsearch==8.17.2
google-cloud-translate==2.0.1
google-genai==1.10.0
grpcio-status==1.49.0rc1
importlib-metadata==8.0.0
ipykernel==6.29.5
ipywidgets==8.1.5
jaraco.collections==5.1.0
langchain==0.3.23
langchain-google-vertexai==2.0.20
minio==7.2.15
nbformat==5.10.4
openpyxl==3.1.5
pandas==2.2.3
pip-chill==1.0.3
plotly==6.0.1
python-dotenv==1.1.0
ruamel.yaml==0.18.10
scikit-learn==1.6.1
tomli==2.0.1
tqdm==4.67.1
xlrd==2.0.1
